+++ frontend/src/pages/WebtoonDetailPage/WebtoonDetailPage.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import WebtoonService from '@services/webtoonService';
 import WebtoonReviewService from '@services/WebtoonReviewService';
-import WebtoonDetailsSection from '@components/WebtoonDetailsSection';
-import WebtoonRatingSection from '@components/WebtoonRatingSection';
-import UserInteractionSection from '@components/UserInteractionSection';
+import WebtoonDetailsSection from '@pages/WebtoonDetailPage/WebtoonDetailsSection';
+import WebtoonRatingSection from '@pages/WebtoonDetailPage/WebtoonRatingSection';
+import UserInteractionSection from '@pages/WebtoonDetailPage/UserInteractionSection';
 import styles from './WebtoonDetailPage.module.css';
 import { Webtoon } from '@models/webtoon';
 import { Review } from '@models/review';
