+++ frontend/src/pages/WebtoonDetailPage/WebtoonRatingSection/index.tsx
